Move risk calculation test helpers to separate extensions
