+++ src/xcode/ENA/ENA/Source/Services/Risk/Calculation/__tests__/RiskCalculationTests.swift
@@ -28,20 +28,20 @@ final class RiskCalculationTests: XCTestCase {
 	// MARK: - Tests for calculating raw risk score
 
 	func testCalculateRawRiskScore_Zero() throws {
-		let summaryZeroMaxRisk = makeExposureSummaryContainer(maxRiskScoreFullRange: 0, ad_low: 10, ad_mid: 10, ad_high: 10)
-		XCTAssertEqual(RiskCalculation.calculateRawRisk(summary: summaryZeroMaxRisk, configuration: appConfig), 0.0, accuracy: 0.01)
+		let summaryZeroMaxRisk: CodableExposureDetectionSummary = .makeExposureSummaryContainer(maxRiskScoreFullRange: 0, ad_low: 10, ad_mid: 10, ad_high: 10)
+		XCTAssertEqual(RiskCalculation.calculateRawRisk(summary: summaryZeroMaxRisk, configuration: .riskCalculationAppConfig), 0.0, accuracy: 0.01)
 	}
 
 	func testCalculateRawRiskScore_Low() throws {
-		XCTAssertEqual(RiskCalculation.calculateRawRisk(summary: summaryLow, configuration: appConfig), 1.07, accuracy: 0.01)
+		XCTAssertEqual(RiskCalculation.calculateRawRisk(summary: .summaryLow, configuration: .riskCalculationAppConfig), 1.07, accuracy: 0.01)
 	}
 
 	func testCalculateRawRiskScore_Med() throws {
-		XCTAssertEqual(RiskCalculation.calculateRawRisk(summary: summaryMed, configuration: appConfig), 2.56, accuracy: 0.01)
+		XCTAssertEqual(RiskCalculation.calculateRawRisk(summary: .summaryMed, configuration: .riskCalculationAppConfig), 2.56, accuracy: 0.01)
 	}
 
 	func testCalculateRawRiskScore_High() throws {
-		XCTAssertEqual(RiskCalculation.calculateRawRisk(summary: summaryHigh, configuration: appConfig), 10.2, accuracy: 0.01)
+		XCTAssertEqual(RiskCalculation.calculateRawRisk(summary: .summaryHigh, configuration: .riskCalculationAppConfig), 10.2, accuracy: 0.01)
 	}
 
 	// MARK: - Tests for calculating risk levels
@@ -55,8 +55,8 @@ final class RiskCalculationTests: XCTestCase {
 			detectionMode: .automatic
 		)
 		let risk = RiskCalculation.risk(
-			summary: summaryHigh,
-			configuration: appConfig,
+			summary: .summaryHigh,
+			configuration: .riskCalculationAppConfig,
 			dateLastExposureDetection: Date(),
 			activeTracing: .init(interval: 48 * 3600),
 			preconditions: preconditions(.invalid),
@@ -82,8 +82,8 @@ final class RiskCalculationTests: XCTestCase {
 		)
 		var risk = RiskCalculation
 			.risk(
-				summary: summaryLow,
-				configuration: appConfig,
+				summary: .summaryLow,
+				configuration: .riskCalculationAppConfig,
 				dateLastExposureDetection: Date(),
 				activeTracing: .init(interval: 0),
 				preconditions: preconditions(.valid),
@@ -96,7 +96,7 @@ final class RiskCalculationTests: XCTestCase {
 		// Test case when summary is nil
 		risk = RiskCalculation.risk(
 			summary: nil,
-			configuration: appConfig,
+			configuration: .riskCalculationAppConfig,
 			dateLastExposureDetection: Date(),
 			activeTracing: .init(interval: 48 * 3600),
 			preconditions: preconditions(.valid),
@@ -118,8 +118,8 @@ final class RiskCalculationTests: XCTestCase {
 
 		// That will happen when the date of last exposure detection is older than one day
 		let risk = RiskCalculation.risk(
-			summary: summaryLow,
-			configuration: appConfig,
+			summary: .summaryLow,
+			configuration: .riskCalculationAppConfig,
 			dateLastExposureDetection: Date().addingTimeInterval(.init(days: -2)),
 			activeTracing: .init(interval: 48 * 3600),
 			preconditions: preconditions(.valid),
@@ -142,7 +142,7 @@ final class RiskCalculationTests: XCTestCase {
 		// That will happen when the date of last exposure detection is older than one day
 		let risk = RiskCalculation.risk(
 			summary: nil,
-			configuration: appConfig,
+			configuration: .riskCalculationAppConfig,
 			dateLastExposureDetection: Date().addingTimeInterval(.init(days: -1)),
 			activeTracing: .init(interval: 48 * 3600),
 			preconditions: preconditions(.valid),
@@ -154,8 +154,8 @@ final class RiskCalculationTests: XCTestCase {
 
 		XCTAssertEqual(
 			RiskCalculation.risk(
-				summary: summaryLow,
-				configuration: appConfig,
+				summary: .summaryLow,
+				configuration: .riskCalculationAppConfig,
 				dateLastExposureDetection: Date().addingTimeInterval(.init(days: -3)),
 				activeTracing: .init(interval: 48 * 3600),
 				preconditions: preconditions(.valid),
@@ -167,8 +167,8 @@ final class RiskCalculationTests: XCTestCase {
 
 		XCTAssertEqual(
 			RiskCalculation.risk(
-				summary: summaryLow,
-				configuration: appConfig,
+				summary: .summaryLow,
+				configuration: .riskCalculationAppConfig,
 				dateLastExposureDetection: Date().addingTimeInterval(.init(days: -1)),
 				activeTracing: .init(interval: 48 * 3600),
 				preconditions: preconditions(.valid),
@@ -189,8 +189,8 @@ final class RiskCalculationTests: XCTestCase {
 
 		// Test the case where preconditions pass and there is low risk
 		let risk = RiskCalculation.risk(
-			summary: summaryLow,
-			configuration: appConfig,
+			summary: .summaryLow,
+			configuration: .riskCalculationAppConfig,
 			// arbitrary, but within limit
 			dateLastExposureDetection: Date().addingTimeInterval(-3600),
 			activeTracing: .init(interval: 48 * 3600),
@@ -209,14 +209,14 @@ final class RiskCalculationTests: XCTestCase {
 			detectionMode: .automatic
 		)
 
-		var summary = summaryHigh
+		var summary: CodableExposureDetectionSummary = .summaryHigh
 		summary.daysSinceLastExposure = 5
 		summary.matchedKeyCount = 10
 
 		// Test the case where preconditions pass and there is increased risk
 		let risk = RiskCalculation.risk(
 			summary: summary,
-			configuration: appConfig,
+			configuration: .riskCalculationAppConfig,
 			// arbitrary, but within limit
 			dateLastExposureDetection: Date().addingTimeInterval(-3600),
 			activeTracing: .init(interval: 48 * 3600),
@@ -240,10 +240,10 @@ final class RiskCalculationTests: XCTestCase {
 
 		// Test the case where preconditions pass and there is increased risk
 		// Values below are hand-picked to result in a raw risk score of 5.12 - within a gap in the range
-		let summary = makeExposureSummaryContainer(maxRiskScoreFullRange: 128, ad_low: 30, ad_mid: 30, ad_high: 30)
+		let summary: CodableExposureDetectionSummary = .makeExposureSummaryContainer(maxRiskScoreFullRange: 128, ad_low: 30, ad_mid: 30, ad_high: 30)
 		let risk = RiskCalculation.risk(
 			summary: summary,
-			configuration: appConfig,
+			configuration: .riskCalculationAppConfig,
 			// arbitrary, but within limit
 			dateLastExposureDetection: Date().addingTimeInterval(-3600),
 			activeTracing: .init(interval: 48 * 3600),
@@ -264,10 +264,10 @@ final class RiskCalculationTests: XCTestCase {
 
 		// Test the case where preconditions pass and there is increased risk
 		// Values below are hand-picked to result in a raw risk score of 13.6 - outside of the top range of the config
-		let summary = makeExposureSummaryContainer(maxRiskScoreFullRange: 255, ad_low: 40, ad_mid: 40, ad_high: 40)
+		let summary: CodableExposureDetectionSummary = .makeExposureSummaryContainer(maxRiskScoreFullRange: 255, ad_low: 40, ad_mid: 40, ad_high: 40)
 		let risk = RiskCalculation.risk(
 			summary: summary,
-			configuration: appConfig,
+			configuration: .riskCalculationAppConfig,
 			// arbitrary, but within limit
 			dateLastExposureDetection: Date().addingTimeInterval(-3600),
 			activeTracing: .init(interval: 48 * 3600),
@@ -288,10 +288,10 @@ final class RiskCalculationTests: XCTestCase {
 
 		// Test the case where preconditions pass and there is increased risk
 		// Values below are hand-picked to result in a raw risk score of 0.85 - outside of the bottom bound of the config
-		let summary = makeExposureSummaryContainer(maxRiskScoreFullRange: 64, ad_low: 10, ad_mid: 10, ad_high: 10)
+		let summary: CodableExposureDetectionSummary = .makeExposureSummaryContainer(maxRiskScoreFullRange: 64, ad_low: 10, ad_mid: 10, ad_high: 10)
 		let risk = RiskCalculation.risk(
 			summary: summary,
-			configuration: appConfig,
+			configuration: .riskCalculationAppConfig,
 			// arbitrary, but within limit
 			dateLastExposureDetection: Date().addingTimeInterval(-3600),
 			activeTracing: .init(interval: 48 * 3600),
@@ -317,8 +317,8 @@ final class RiskCalculationTests: XCTestCase {
 		// Test case where last exposure summary was gotten too far in the past,
 		// But the risk is increased
 		let risk = RiskCalculation.risk(
-			summary: summaryHigh,
-			configuration: appConfig,
+			summary: .summaryHigh,
+			configuration: .riskCalculationAppConfig,
 			dateLastExposureDetection: Date().addingTimeInterval(.init(days: -2)),
 			activeTracing: .init(interval: 48 * 3600),
 			preconditions: preconditions(.valid),
@@ -341,8 +341,8 @@ final class RiskCalculationTests: XCTestCase {
 		// Active tracing is more then 24h
 		// But the risk is increased
 		let risk = RiskCalculation.risk(
-			summary: summaryHigh,
-			configuration: appConfig,
+			summary: .summaryHigh,
+			configuration: .riskCalculationAppConfig,
 			dateLastExposureDetection: Date().addingTimeInterval(.init(hours: -23)),
 			activeTracing: .init(interval: 48 * 3600),
 			preconditions: preconditions(.valid),
@@ -365,8 +365,8 @@ final class RiskCalculationTests: XCTestCase {
 		// Active tracing is less then 24h
 		// But the risk is increased
 		let risk = RiskCalculation.risk(
-			summary: summaryHigh,
-			configuration: appConfig,
+			summary: .summaryHigh,
+			configuration: .riskCalculationAppConfig,
 			dateLastExposureDetection: Date().addingTimeInterval(.init(hours: -25)),
 			activeTracing: .init(interval: 2 * 3600),
 			preconditions: preconditions(.valid),
@@ -389,8 +389,8 @@ final class RiskCalculationTests: XCTestCase {
 		// Active tracing is less then 24h
 		// But the risk is increased
 		let risk = RiskCalculation.risk(
-			summary: summaryHigh,
-			configuration: appConfig,
+			summary: .summaryHigh,
+			configuration: .riskCalculationAppConfig,
 			dateLastExposureDetection: Date().addingTimeInterval(.init(hours: -23)),
 			activeTracing: .init(interval: 2 * 3600),
 			preconditions: preconditions(.valid),
@@ -412,7 +412,7 @@ final class RiskCalculationTests: XCTestCase {
 		// Test case where last exposure summary was gotten too far in the past,
 		let risk = RiskCalculation.risk(
 			summary: nil,
-			configuration: appConfig,
+			configuration: .riskCalculationAppConfig,
 			dateLastExposureDetection: Date().addingTimeInterval(.init(days: -2)),
 			activeTracing: .init(interval: 48 * 3600),
 			preconditions: preconditions(.valid),
@@ -437,8 +437,8 @@ final class RiskCalculationTests: XCTestCase {
 
 		// Will produce increased risk
 		let risk = RiskCalculation.risk(
-			summary: summaryHigh,
-			configuration: appConfig,
+			summary: .summaryHigh,
+			configuration: .riskCalculationAppConfig,
 			// arbitrary, but within limit
 			dateLastExposureDetection: Date().addingTimeInterval(-3600),
 			activeTracing: .init(interval: 48 * 3600),
@@ -463,8 +463,8 @@ final class RiskCalculationTests: XCTestCase {
 
 		// Will produce high risk
 		let risk = RiskCalculation.risk(
-			summary: summaryHigh,
-			configuration: appConfig,
+			summary: .summaryHigh,
+			configuration: .riskCalculationAppConfig,
 			// arbitrary, but within limit
 			dateLastExposureDetection: Date().addingTimeInterval(-3600),
 			activeTracing: .init(interval: 48 * 3600),
@@ -487,8 +487,8 @@ final class RiskCalculationTests: XCTestCase {
 		// and the new risk level has not changed
 
 		let risk = RiskCalculation.risk(
-			summary: summaryLow,
-			configuration: appConfig,
+			summary: .summaryLow,
+			configuration: .riskCalculationAppConfig,
 			// arbitrary, but within limit
 			dateLastExposureDetection: Date().addingTimeInterval(-3600),
 			activeTracing: .init(interval: 48 * 3600),
@@ -512,8 +512,8 @@ final class RiskCalculationTests: XCTestCase {
 
 		// Produces unknown risk
 		let risk = RiskCalculation.risk(
-			summary: summaryLow,
-			configuration: appConfig,
+			summary: .summaryLow,
+			configuration: .riskCalculationAppConfig,
 			dateLastExposureDetection: Date().addingTimeInterval(.init(days: -2)),
 			activeTracing: .init(interval: 48 * 3600),
 			preconditions: preconditions(.valid),
@@ -536,8 +536,8 @@ final class RiskCalculationTests: XCTestCase {
 
 		// Produces unknown risk
 		let risk = RiskCalculation.risk(
-			summary: summaryLow,
-			configuration: appConfig,
+			summary: .summaryLow,
+			configuration: .riskCalculationAppConfig,
 			dateLastExposureDetection: Date().addingTimeInterval(.init(days: -2)),
 			activeTracing: .init(interval: 48 * 3600),
 			preconditions: preconditions(.valid),
@@ -552,22 +552,6 @@ final class RiskCalculationTests: XCTestCase {
 // MARK: - Helpers
 
 private extension RiskCalculationTests {
-
-	private var appConfig: SAP_ApplicationConfiguration {
-		makeAppConfig(w_low: 1.0, w_med: 0.5, w_high: 0.5)
-	}
-
-	private var summaryLow: CodableExposureDetectionSummary {
-		makeExposureSummaryContainer(maxRiskScoreFullRange: 80, ad_low: 10, ad_mid: 10, ad_high: 10)
-	}
-
-	private var summaryMed: CodableExposureDetectionSummary {
-		makeExposureSummaryContainer(maxRiskScoreFullRange: 128, ad_low: 15, ad_mid: 15, ad_high: 15)
-	}
-
-	private var summaryHigh: CodableExposureDetectionSummary {
-		makeExposureSummaryContainer(maxRiskScoreFullRange: 255, ad_low: 30, ad_mid: 30, ad_high: 30)
-	}
 
 	enum PreconditionState {
 		case valid
@@ -588,56 +572,4 @@ private extension RiskCalculationTests {
 		}
 	}
 
-	private func makeExposureSummaryContainer(
-		maxRiskScoreFullRange: Int,
-		ad_low: Double,
-		ad_mid: Double,
-		ad_high: Double
-	) -> CodableExposureDetectionSummary {
-		.init(
-			daysSinceLastExposure: 0,
-			matchedKeyCount: 0,
-			maximumRiskScore: 0,
-			attenuationDurations: [ad_low, ad_mid, ad_high],
-			maximumRiskScoreFullRange: maxRiskScoreFullRange
-		)
-	}
-
-	/// Makes an mock `SAP_ApplicationConfiguration`
-	///
-	/// Some defaults are applied for ad_norm, w4, and low & high ranges
-	private func makeAppConfig(
-		ad_norm: Int32 = 25,
-		w4: Int32 = 0,
-		w_low: Double,
-		w_med: Double,
-		w_high: Double,
-		riskRangeLow: ClosedRange<Int32> = 1...5,
-		// Gap between the ranges is on purpose, this is an edge case to test
-		riskRangeHigh: Range<Int32> = 6..<11
-	) -> SAP_ApplicationConfiguration {
-		var config = SAP_ApplicationConfiguration()
-		config.attenuationDuration.defaultBucketOffset = w4
-		config.attenuationDuration.riskScoreNormalizationDivisor = ad_norm
-		config.attenuationDuration.weights.low = w_low
-		config.attenuationDuration.weights.mid = w_med
-		config.attenuationDuration.weights.high = w_high
-
-		var riskScoreClassLow = SAP_RiskScoreClass()
-		riskScoreClassLow.label = "LOW"
-		riskScoreClassLow.min = riskRangeLow.lowerBound
-		riskScoreClassLow.max = riskRangeLow.upperBound
-
-		var riskScoreClassHigh = SAP_RiskScoreClass()
-		riskScoreClassHigh.label = "HIGH"
-		riskScoreClassHigh.min = riskRangeHigh.lowerBound
-		riskScoreClassHigh.max = riskRangeHigh.upperBound
-
-		config.riskScoreClasses.riskClasses = [
-			riskScoreClassLow,
-			riskScoreClassHigh
-		]
-
-		return config
-	}
 }
